+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Only one connection is opened at a time.
 - Randomized connection selection when multiple one is available.
   (Provides better distribution of queries, less overuse of single connection thread.)
+- Generated `==` and `hashCode` for `PgUrl`.
 
 ## 1.0.0